requirements.txt
# Core dependencies for Real-time Data Pipeline

# Apache Beam and Dataflow
apache-beam[gcp]==2.52.0

# Google Cloud Platform
google-cloud-pubsub==2.19.0
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-auth==2.25.2
google-cloud-logging==3.5.0
google-cloud-monitoring==2.15.0

# Apache Airflow
apache-airflow-providers-google==10.12.0
apache-airflow-providers-apache-beam==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.1

# Configuration and utilities
PyYAML==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0


# Monitoring and observability
prometheus-client==0.19.0
statsd==4.0.1

# Testing and development
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-xdist==3.5.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Security and validation
cryptography==41.0.8
jsonschema==4.20.0
cerberus==1.3.5

# Performance and optimization
ujson==5.8.0
orjson==3.9.10

# Logging and debugging
structlog==23.2.0
colorlog==6.8.0

# HTTP and networking
requests==2.31.0
httpx==0.25.2
urllib3==2.1.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Database connectors (if needed for additional sources)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Task scheduling (alternative to Airflow)
celery==5.3.4
redis==5.0.1

# Container and deployment
docker==6.1.3
kubernetes==28.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development tools
pre-commit==3.6.0
jupyter==1.0.0
ipython==8.17.2

# Version pinning for stability
setuptools==69.0.3
wheel==0.42.0
pip==23.3.2

# Optional: Additional monitoring and alerting
sentry-sdk==1.38.0
slack-sdk==3.23.0

# Optional: Advanced analytics
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Optional: Stream processing alternatives
faust-streaming==0.10.4
confluent-kafka==2.3.0

# Security scanning
bandit==1.7.5
safety==2.3.5

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0
